Generic access token creation implemented.

+++ src/storage/user_tokens/scylla_user_token_repository.rs
@@ -60,7 +60,7 @@ impl UserTokenRepository for ScyllaUserTokenRepository {
         )).await?;
 
         let rows = result.rows.unwrap(); 
-        let row = rows.into_typed::<(i32,)>().next().unwrap();
+        let row = rows.into_typed::<(i64,)>().next().unwrap();
         let (success, ) = row?;
         Ok(success > 0)
     }
